Created a better input system making try_until_in and related functions obsolete.
System uses an optional parser that attempts to change the type of an
input (restarting the input if unavailable) and an optional validator to
check other conditions such as try_until_in_set. added custom type
aliases to make the functions required to fullfil the validator and
parser roles clearer.